simplify init nuxt instance from preview
