+++ packages/storybook-nuxt/src/preview.ts
@@ -1,19 +1,44 @@
 import consola from 'consola'
+import { applyPlugins, createNuxtApp } from 'nuxt/app'
+import { nextTick } from 'vue'
 
-const vueAppRootContainer = document.createElement('div')
-vueAppRootContainer.id = '__nuxt'
-vueAppRootContainer.setAttribute('hidden', 'true')
-document.body.appendChild(vueAppRootContainer)
+// @ts-expect-error virtual file
+import plugins from '#build/plugins'
 
-async function nuxtAppEntry() {
-  const nuxtApp = () => import(('#app/entry')).then(m => m.default).catch(() => {})
-  // i
-  const vueAppPromise = nuxtApp().catch((error) => {
-    consola.error('Error while mounting app:', error)
-  })
-  return vueAppPromise
+import '#build/css'
+
+const globalWindow = window as any
+
+async function applyNuxtPlugins(vueApp: any, storyContext: any) {
+  const nuxt = createNuxtApp({ vueApp, globalName: `nuxt-${storyContext.id}` })
+
+  async function handleVueError(err: any) {
+    await nuxt.callHook('app:error', err)
+    nuxt.payload.error = (nuxt.payload.error || err) as any
+  }
+
+  vueApp.config.errorHandler = handleVueError
+
+  try {
+    await applyPlugins(nuxt, plugins)
+  }
+  catch (err) {
+    handleVueError(err)
+  }
+  try {
+    await nuxt.hooks.callHook('app:created', vueApp)
+    await nuxt.hooks.callHook('app:beforeMount', vueApp)
+    setTimeout(async () => {
+      await nuxt.hooks.callHook('app:mounted', vueApp)
+      await nextTick()
+    }, 10)
+  }
+  catch (e) {
+    consola.error('Vue Error in plugins ', e)
+  }
+
+  return nuxt
 }
 
-nuxtAppEntry().then((app) => {
-  consola.info('nuxtAppEntry done', app)
-})
+globalWindow.PLUGINS_SETUP_FUNCTIONS ||= new Set()
+globalWindow.PLUGINS_SETUP_FUNCTIONS.add(applyNuxtPlugins)
